nvme effects-log block device is not allowed
When using the "effects-log" command in nvme-cli without specifying the "csi" argument,
a default action is added to check the ctrl register,
allowing only character devices like nvme0.

In older versions of nvme-cli, the default value of csi was set to 0 and issued without any problems
But it has been fixed now.
It can be used regardless of the version when used as a character device.

Reference:
https://github.com/linux-nvme/nvme-cli/pull/2742

Signed-off-by: Steven Seungcheol Lee <sc108.lee@samsung.com>